+++ systems/__init__.py
@@ -1,1 +1,0 @@
-from .LCNJP import LCNJP