+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+## 2021/05/16 - 00 - PRL Model
+> Toolbox version 0.7.0
+* Updated and renamed `LCNJP` model to `NJP00`.
+* Added `PRL00` for Phys. Rev. Lett. **103**, 213603.
+* Added script for Figure 2b of Phys. Rev. Lett. **103**, 213603.
+
 ## 2021/05/14 - 00 - NJP Model
 > Toolbox version 0.7.0
-* Added scripts for Figure 2 and 3a of New J. Phys. 22, 063041.
-* Added `LCNJP` system for New J. Phys. 22, 063041.
+* Added scripts for Figure 2 and 3a of New J. Phys. **22**, 063041.
+* Added `LCNJP` system for New J. Phys. **22**, 063041.
 * Updated `CHANGELOG`, `GUI` and `README`.